fix(module8-task1): changes after reviews
